Cree archivos independientes para todas las clases
