+++ objects_class/actress.rb
@@ -1,0 +1,13 @@
+class Actress < Entertainment
+  def initialize(netflix_show, genre, schedule, price)
+    @netflix_show = netflix_show
+    @genre = genre
+    @schedule = schedule
+    @price = price
+  end
+
+  #show en el que sale
+  def actress_shows
+    "Show 1"
+  end
+end
