+++ objects_class/entertaiment.rb
@@ -1,0 +1,3 @@
+class Entertainment
+  #relacionar tv con actriz
+end
